fix: rename interfaces to ts
